Algorithm_evaluation/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
astor=0.8.1=pypi_0
blas=1.0=mkl
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.9.24=py39h06a4308_0
charset-normalizer=2.1.1=pypi_0
cloudpickle=2.2.0=pypi_0
colorama=0.4.6=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.0.6=pypi_0
cycler=0.11.0=pypi_0
dnspython=2.2.1=pypi_0
einops=0.6.0=pypi_0
filelock=3.8.2=pypi_0
fonttools=4.38.0=pypi_0
future=0.18.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-auth=2.15.0=pypi_0
grpcio=1.51.1=pypi_0
h5py=3.7.0=py39h737f45e_0
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.12.0=pypi_0
hyperopt=0.1.2=pypi_0
idna=3.4=pypi_0
imageio=2.25.0=pypi_0
importlib-metadata=5.1.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pypi_0
json-tricks=3.16.1=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.2=pypi_0
mkl-service=2.4.0=py39h7f8727e_0
mkl=2021.4.0=h06a4308_640
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
ncurses=6.3=h5eee18b_3
networkx=2.8.8=pypi_0
nni=2.5=pypi_0
numpy-base=1.23.5=py39h31eccc5_0
numpy=1.23.5=py39h14f4228_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.68=pypi_0
openssl=1.1.1s=h7f8727e_0
packaging=22.0=pypi_0
pandas=1.5.2=pypi_0
pillow=9.3.0=pypi_0
pip=22.3.1=py39h06a4308_0
prettytable=3.5.0=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyasn1=0.4.8=pypi_0
pymongo=4.3.3=pypi_0
pyparsing=3.0.9=pypi_0
python-dateutil=2.8.2=pypi_0
python=3.9.15=h7a1cb2a_2
pythonwebhdfs=0.2.3=pypi_0
pytz=2022.6=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests-oauthlib=1.3.1=pypi_0
requests=2.28.1=pypi_0
responses=0.22.0=pypi_0
rsa=4.9=pypi_0
schema=0.7.5=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.2.0=pypi_0
scipy=1.9.3=pypi_0
setuptools=59.5.0=pypi_0
simplejson=3.18.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
spikingjelly=0.0.0.0.12=pypi_0
sqlite=3.40.0=h5082296_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboard=2.11.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.1.23.1=pypi_0
timm=0.6.12=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
torch=1.10.1+cu111=pypi_0
torchaudio=0.10.1+cu102=pypi_0
torchinfo=1.7.1=pypi_0
torchvision=0.11.2+cu111=pypi_0
tqdm=4.64.1=pypi_0
types-toml=0.10.8.1=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=pypi_0
wcwidth=0.2.5=pypi_0
websockets=10.4=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.8=h5eee18b_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
